% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model2netcdf.CABLE.R
\name{model2netcdf.CABLE}
\alias{model2netcdf.CABLE}
\title{Code to convert CABLE's output into netCDF format}
\usage{
model2netcdf.CABLE(outdir, sitelat, sitelon, start_date, end_date)
}
\arguments{
\item{outdir}{Location of model output}

\item{sitelat}{Latitude of the site}

\item{sitelon}{Longitude of the site}

\item{start_date}{Start time of the simulation}

\item{end_date}{End time of the simulation}
}
\description{
Convert CABLE output into the NACP Intercomparison format (ALMA using netCDF)
}
\author{
Rob Kooper
}
