% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write.config.CABLE.R
\name{write.config.CABLE}
\alias{write.config.CABLE}
\title{Write CABLE configuration files}
\usage{
write.config.CABLE(defaults, trait.values, settings, run.id)
}
\arguments{
\item{defaults}{list of defaults to process}

\item{settings}{list of settings from pecan settings file}

\item{run.id}{id of run}

\item{trait.samples}{vector of samples for a given trait}
}
\value{
configuration file for CABLE for given run
}
\description{
Writes a CABLE config file.
}
\details{
Requires a pft xml object, a list of trait values for a single model run,
and the name of the file to create
}
\author{
Rob Kooper, Kaitlin Ragosta
}
