% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_restart.CABLE.R
\name{write_restart.CABLE}
\alias{write_restart.CABLE}
\title{Write restart template for SDA}
\usage{
write_restart.CABLE(outdir, runid, start.time, stop.time, settings, new.state)
}
\arguments{
\item{start.time}{Time of current assimilation step}

\item{stop.time}{Time of next assimilation step}

\item{new.state}{Analysis state matrix returned by \code{sda.enkf}}
}
\description{
Write restart files for model
}
\author{
Alexey Shiklomanov
}
